requirements.txt
Flask>=1.0.3
gunicorn>=19.9.0
requests
pillow
beautifulsoup4
zalgoify>=1.0
upsidedown
praw